Changed some stuff but frames be off
Changed some stuff but the frames seem to be slightly off and could be responsible for the error in position when trying to navigate to the placement pose (given in map frame). Additionally, ODOM might be map frame? Also, difference in base_footprint to base_link and base_cover_link might be responsible for weak grip (gripping only upper part of the aruco "cube")