Fixed C part of lab with behaviour tree
C part of lab now done. Interestingly the robot does seem to get a much better grasp of the "cube" most of the time when using sensor instead. Perhaps the original pose given in the launch file is a bit off? Additionally sometimes robot veers off-course when going to rotate and then drive to other table. Don't know why this is though.